api-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
aiohttp==3.8.6
python-multipart==0.0.6
python-dotenv==1.0.0
yt-dlp
uvloop